impl benchmark api
