+++ pallets/tendermint-light-client/src/benchmarks.rs
@@ -1,32 +1,41 @@
 use super::*;
-use crate::{mock, types::*, Pallet as TendermintLightClient};
+
+// use crate::{mock, types, Pallet as TendermintLightClient};
+use crate::Pallet as TendermintLightClient;
 use frame_benchmarking::{account, benchmarks, benchmarks_instance_pallet, whitelisted_caller};
 use frame_system::{Origin, RawOrigin};
 use sp_runtime::traits::Bounded;
-use std::mem;
 
 benchmarks! {
     benchmark_for_initialize_client {
 
+        let i in 0 .. i32::max_value() as u32;
+
         let caller = RawOrigin::Root; //whitelisted_caller();
-        let options = LightClientOptionsStorage::default ();
+        let options = types::LightClientOptionsStorage::default ();
 
-        // TODO : for all possible memory sizes of struct
-        // TODO : benchmark for all Vec<T>
-        // let size_ = mem::size_of::<LightBlockStorage>();
-        let initial_block: LightBlockStorage = serde_json::from_str(mock::TRUSTED_BLOCK).unwrap();
+        // let initial_block = mock::create_block (i as i32, i as i32, i as i32, i as i32);
 
-    }: initialize_client(caller.clone(), options, initial_block)
+        let initial_block = types::LightBlockStorage::create (i as i32, i as i32, i as i32, i as i32);
+
+    }: initialize_client(caller.clone(), options, initial_block.clone ())
 
     verify {
-        // check if event fired
-        assert_eq! (true, true);
+        // check if persisted
+        let last_imported = TendermintLightClient::<T>::get_last_imported_hash();
+        assert_eq!(
+            initial_block.signed_header.commit.block_id.hash,
+            last_imported
+        );
+
+        // assert_eq! (true, true);
 
     }
 
-    // impl_benchmark_test_suite!(
-    //     TendermintLightClient,
-    //     crate::mock::ExtBuilder::default().existential_deposit(256).build(),
-    //     crate::mock::Test,
-    // );
+    impl_benchmark_test_suite!(
+        TendermintLightClient,
+        mock::new_genesis_storage (),
+        mock::TestRuntime,
+    );
+
 }
